Use Luminus to create a new project
Goal:
This is my first Clojure web app and first real app with any kind of
compelxity potential behind it. As such I wanted to use Luminus which is
a template with defaults for web apps that a lot of people use. My goal
with v0 is to supply an api for Pathfinder spells that returns all the
info about a spell with a JSON schema compliant rest payload. For V0 the
only client will be an amazon alexa skill

Changes:
- ran `lein new luminus +postgres +swagger`
- Skipped clojurescript setup for now as that will be easy to add and
not a v0 (api only) priority